Downgrade cucumber version to align them with cucumber-android dependencies
