+++ android/src/androidTest/assets/features/Login.feature
@@ -1,11 +1,11 @@
 Feature: Login
   Scenario: Login screen
-    Given I am logged "out"
-    Then I see the "Login" screen
-    Then I see the "Email" field
-    Then I type a random email in the "Email" field
-    Then I see the "Password" field
-    Then I type a random text in the "Password" field
+    Given I am in the "Login" screen
+    Then I see the "Email" textfield with text "Email"
+    Then I type "test@test.com" in the email field
+    Then I see the "Password" textfield with text "Password"
+    Then I type "1234" in the password field
     Then I see the "Login" button
-    Then I press the "Login" button
-    Then I see the "Details" screen
+    Then I press the login button
+    Then I see the "Home" screen
+    Then I see "test@test.com" text
